requirements.txt
dj-database-url==1.0.0
Django==4.2.13
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
tzdata==2022.2
boto3==1.26.41
botocore==1.29.41
django-storages==1.13.2
django-import-export==3.0.2
pillow==9.4.0
antlr4-python3-runtime==4.11.0
sympy==1.12

certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
requests==2.31.0
beautifulsoup4==4.12.2
soupsieve==2.4.1

channels==3.0.5
channels-redis==4.2.0
# daphne==3.0.0
